--- a/lib/enum.rb
+++ b/lib/enum.rb
@@ -7,13 +7,23 @@ class Enum
   alias_method :inspect, :name
 
   def initialize(name, value)
+    @short_name = name.to_s.underscore
     @name = "#{self.class.name}::#{name}"
-    @value = value || @name
+    @value = value
   end
 
   class << self
+    attr_accessor :members
+
+    def inherited(child)
+      child.instance_eval do
+        @values = {}
+        @members = []
+      end
+    end
+
     def method_missing(name, *args, **kwargs, &block)
-      return super unless name[0] == name[0].upcase
+      return super unless name[0] =~ /[A-Z]/
       new(name, *args, **kwargs, &block)
     end
 
@@ -22,21 +32,16 @@ class Enum
     end
 
     def values
-      map(&:value)
+      @values.keys
     end
 
     def each(&block)
-      @members.each(&block) if defined?(@members)
-    end
-
-    def members
-      return @members if defined?(@members)
-      @members = []
+      @members.each(&block)
     end
 
     private
 
-    def new(name, value = nil, &block)
+    def new(name, value = name, &block)
       if self == Enum
         raise ArgumentError,
           "You can't add values to the abstract Enum class itself."
@@ -47,25 +52,24 @@ class Enum
           "Name conflict: '#{self.name}::#{name}' is already defined."
       end
 
-      if values.include?(value)
+      if @values[value]
         raise ArgumentError,
           "Value conflict: the value '#{value}' is defined for '#{self.cast(value).name}'."
       end
 
-      member = super(name, value)
-      member.instance_eval(&block) if block_given?
-
       class_eval(<<-RUBY, __FILE__, __LINE__ + 1)
         def #{name.to_s.underscore}?
-          name.demodulize.underscore == "#{name.to_s.underscore}"
+          @short_name == "#{name.to_s.underscore}"
         end
       RUBY
 
+      member = super(name, value)
+      member.instance_eval(&block) if block_given?
       member.freeze
 
       const_set(name, member)
-      (@members ||= []) << member
-      (@values ||= {})[value] = member
+      @members << member
+      @values[value] = member
     end
   end
 end