--- a/lib/literal_enums/transitions.rb
+++ b/lib/literal_enums/transitions.rb
@@ -1,32 +1,48 @@
 module LiteralEnums
   module Transitions
     def respond_to_missing?(name, include_private = false)
-      return false if name == :transitions_to
-      valid_next_states.any? { |m| m.name.to_s.demodulize.underscore.to_sym == name }
+      case transitions_to
+      when Enum
+        transitions_to.name.to_s.demodulize.underscore.to_sym == name
+      when Array
+        transitions_to.any? { |m| m.name.to_s.demodulize.underscore.to_sym == name }
+      end
     end
 
     def method_missing(name, *args, **kwargs, &block)
-      valid_next_states.find { |m| m.name.to_s.demodulize.underscore.to_sym == name } || super
+      case transitions_to
+      when Enum
+        if transitions_to.name.to_s.demodulize.underscore.to_sym == name
+          return transitions_to
+        end
+      when Array
+        if (next_one = transitions_to.find { |m| m.name.to_s.demodulize.underscore.to_sym == name })
+          return next_one
+        end
+      end
+
+      super
     end
 
     def transition_to(new_state)
       return new_state if can_transition_to?(new_state)
 
-      raise TransitionError, "You can't transition from #{self.name} to #{new_state.name}."
+      raise TransitionError,
+        "You can't transition from #{self.name} to #{new_state.name}."
     end
 
     def can_transition_to?(new_state)
-      unless new_state.is_a?(self.class)
-        raise ArgumentError "You can only transition to another #{self.class.name}."
+      case transitions_to
+      when Enum
+        transitions_to == new_state
+      when Array
+        transitions_to.include? new_state
       end
-
-      valid_next_states.include?(new_state)
     end
 
     private
 
-    def valid_next_states
-      return Array(transitions_to) if respond_to? :transitions_to
+    def transitions_to
       []
     end
   end