Before
```
                Cast    518.327k (± 0.2%) i/s -      2.598M in   5.012406s
              Lookup     23.886M (± 0.6%) i/s -    120.990M in   5.065415s
             Members    775.036k (± 0.3%) i/s -      3.939M in   5.082299s
              Values    374.491k (± 0.2%) i/s -      1.905M in   5.085853s
```


After
```
                Cast     12.079M (± 0.2%) i/s -     61.547M in   5.095414s
              Lookup     24.132M (± 0.5%) i/s -    120.758M in   5.004239s
             Members     17.562M (± 0.4%) i/s -     87.971M in   5.009275s
              Values     10.780M (± 0.2%) i/s -     54.922M in   5.094745s
```